finally got ships rendered at least partially after redux drama

+++ src/actions/single/FetchActions.js
@@ -20,7 +20,7 @@ export const fetchGames = () => {
 
 // This isn't being called
 export const unmountFetchGames = () => {
-	fetchGamesQuery.firestore.INTERNAL.delete();
+	if(fetchGamesQuery.firestore.INTERNAL.delete) fetchGamesQuery.firestore.INTERNAL.delete();
 	console.log('unmounting query');
 	return{
 		type: Types.UNMOUNT_FETCH_GAMES
@@ -29,16 +29,23 @@ export const unmountFetchGames = () => {
 
 let fetchGameQuery;
 
-export const fetchGame = (uid) => {
+export const fetchGame = (_id) => {
+	if(typeof fetchGameQuery === 'function'){
+		console.log('undoing');
+		fetchGamesQuery();
+	}
 	return dispatch => {
 		dispatch({type: Types.FETCH_GAME});
-		fetchGameQuery = firebase.firestore().collection('single').where('userId', '==', uid).onSnapshot(snapshot => {
+		fetchGameQuery = firebase.firestore().collection('single').where('_id', '==', _id).onSnapshot(snapshot => {
+			console.log(snapshot.docs[0], 'should be game returned on fetch');
 			dispatch({type: Types.FETCHED_GAME, payload: snapshot.docs[0].data()});
 		});
 	}
 };
 
 export const unmountFetchGame = () => {
-	fetchGameQuery();
+	if(typeof fetchGameQuery === 'function'){
+		fetchGameQuery();
+	}
 	return{type: 'bla'};
 };
